requirements_linux.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Pillow>=10.0.0
reportlab==4.0.7
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
email-validator==2.1.0
werkzeug==3.0.1
pandas>=2.0.0
python-docx==1.1.0
PyPDF2==3.0.1
pycryptodome==3.23.0
